+++ src/theme/config/typography.ts
@@ -1,26 +1,31 @@
-/*
-* Contents of this configuration have been generated with [Mui Figma Plugin](https://stash.grupa.onet/projects/RDS/repos/mui-figma-plugin/browse)
-* Do not modify these values directly.
-*/
 import { TypographyOptions } from '@mui/material/styles/createTypography';
 
 const HTML_FONT_SIZE = 16;
+const FONT_FAMILY = 'Arial, sans-serif';
 
 type ExtendedTypographyOptions = TypographyOptions
     & {pxToRem: (size: number) => string;};
 
 export const typography: ExtendedTypographyOptions = {
     htmlFontSize: HTML_FONT_SIZE,
+    fontFamily: FONT_FAMILY,
+    button: {
+        textTransform: 'uppercase',
+        fontSize: '14px',
+        fontFamily: FONT_FAMILY,
+        fontWeight: 600,
+        fontStyle: 'normal'
+    },
     h1: {
         fontSize: '96px',
-        fontFamily: 'Neue Haas Grotesk Text Pro, sans-serif',
+        fontFamily: FONT_FAMILY,
         lineHeight: '112px',
         fontWeight: 400,
         fontStyle: 'normal'
     },
     h2: {
         fontSize: '60px',
-        fontFamily: 'Neue Haas Grotesk Text Pro, sans-serif',
+        fontFamily: FONT_FAMILY,
         letterSpacing: '-0.5px',
         lineHeight: '72px',
         fontWeight: 400,
@@ -28,29 +33,29 @@ export const typography: ExtendedTypographyOptions = {
     },
     h3: {
         fontSize: '48px',
-        fontFamily: 'Neue Haas Grotesk Text Pro, sans-serif',
+        fontFamily: FONT_FAMILY,
         lineHeight: '56px'
     },
     h4: {
         fontSize: '34px',
-        fontFamily: 'Neue Haas Grotesk Text Pro, sans-serif',
+        fontFamily: FONT_FAMILY,
         lineHeight: '42px'
     },
     h5: {
         fontSize: '22px',
-        fontFamily: 'Neue Haas Grotesk Text Pro, sans-serif',
+        fontFamily: FONT_FAMILY,
         lineHeight: '32px'
     },
     h6: {
         fontSize: '18px',
-        fontFamily: 'Neue Haas Grotesk Text Pro, sans-serif',
+        fontFamily: FONT_FAMILY,
         lineHeight: '24px',
         fontWeight: 600,
         fontStyle: 'normal'
     },
     body1: {
         fontSize: '14px',
-        fontFamily: 'Neue Haas Grotesk Text Pro, sans-serif',
+        fontFamily: FONT_FAMILY,
         letterSpacing: '0.15px',
         lineHeight: '24px',
         fontWeight: 400,
@@ -58,14 +63,14 @@ export const typography: ExtendedTypographyOptions = {
     },
     body2: {
         fontSize: '13px',
-        fontFamily: 'Neue Haas Grotesk Text Pro, sans-serif',
+        fontFamily: FONT_FAMILY,
         lineHeight: '20px',
         fontWeight: 400,
         fontStyle: 'normal'
     },
     subtitle1: {
         fontSize: '16px',
-        fontFamily: 'Neue Haas Grotesk Text Pro, sans-serif',
+        fontFamily: FONT_FAMILY,
         letterSpacing: '0.15px',
         lineHeight: '28px',
         fontWeight: 400,
@@ -73,14 +78,14 @@ export const typography: ExtendedTypographyOptions = {
     },
     subtitle2: {
         fontSize: '14px',
-        fontFamily: 'Neue Haas Grotesk Text Pro, sans-serif',
+        fontFamily: FONT_FAMILY,
         lineHeight: '22px',
         fontWeight: 600,
         fontStyle: 'normal'
     },
     overline: {
         fontSize: '12px',
-        fontFamily: 'Neue Haas Grotesk Text Pro, sans-serif',
+        fontFamily: FONT_FAMILY,
         letterSpacing: '1px',
         lineHeight: '36px',
         textTransform: 'uppercase',
@@ -89,14 +94,14 @@ export const typography: ExtendedTypographyOptions = {
     },
     caption: {
         fontSize: '12px',
-        fontFamily: 'Neue Haas Grotesk Text Pro, sans-serif',
-        lineHeight: '20px',
+        fontFamily: FONT_FAMILY,
+        lineHeight: '18px',
         fontWeight: 400,
         fontStyle: 'normal'
     },
     label: {
         fontSize: '12px',
-        fontFamily: 'Neue Haas Grotesk Text Pro, sans-serif',
+        fontFamily: FONT_FAMILY,
         lineHeight: '20px',
         textTransform: 'uppercase',
         fontWeight: 400,
@@ -104,7 +109,7 @@ export const typography: ExtendedTypographyOptions = {
     },
     headline1: {
         fontSize: '16px',
-        fontFamily: 'Neue Haas Grotesk Text Pro, sans-serif',
+        fontFamily: FONT_FAMILY,
         lineHeight: '28px',
         textTransform: 'uppercase',
         fontWeight: 600,
@@ -112,7 +117,7 @@ export const typography: ExtendedTypographyOptions = {
     },
     headline2: {
         fontSize: '14px',
-        fontFamily: 'Neue Haas Grotesk Text Pro, sans-serif',
+        fontFamily: FONT_FAMILY,
         lineHeight: '24px',
         textTransform: 'uppercase',
         fontWeight: 600,
@@ -120,7 +125,7 @@ export const typography: ExtendedTypographyOptions = {
     },
     headline3: {
         fontSize: '12px',
-        fontFamily: 'Neue Haas Grotesk Text Pro, sans-serif',
+        fontFamily: FONT_FAMILY,
         lineHeight: '22px',
         textTransform: 'uppercase',
         fontWeight: 600,