+++ src/theme/theme.tsx
@@ -129,6 +129,11 @@ export const getTheme = (
                             },
                             '.MuiInputBase-root': {
                                 borderColor: 'red!important'
+                            },
+                            '.MuiFormLabel-root': {
+                                '&.Mui-focused, &.MuiInputLabel-shrink': {
+                                    transform: 'translate(0, 1.5px) scale(0.9)'
+                                }
                             }
                         };
                     }